+++ internal/entities/permission-entity.go
@@ -1,15 +1,13 @@
 package entities
+
 import (
 	"request-system/pkg/types"
 )
+
 type Permission struct {
-	ID          int    `json:"id"`
+	ID          uint64 `json:"id"`
 	Name        string `json:"name"`
 	Description string `json:"description"`
 
-	types.BaseEntity 
+	types.BaseEntity
 }
-
-
-
-
